# swagger.yaml
openapi: 3.0.0
info:
  title: Hệ Thống Bán Hàng API
  version: 1.0.0
  description: Tài liệu cho 15 API (Users, Products, Categories)
servers:
  - url: http://localhost:3000
    description: Local Server

# --- ĐỊNH NGHĨA MODEL DỮ LIỆU ---
components:
  schemas:
    User:
      type: object
      required:
        - username
        - password
        - email
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        phone:
          type: string
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, inactive]
...
paths:
...
  /users/register:
    post:
      tags: [Users]
      summary: 1. Đăng ký tài khoản mới (Create)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        200:
          description: Đăng ký thành công
  /users/login:
    post:
      tags: [Users]
      summary: 2. Đăng nhập
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: Đăng nhập thành công (Trả về Token)
...
  /users/{id}:
    get:
      tags: [Users]
      summary: 4. Lấy chi tiết người dùng (Read One)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Thông tin user
    
    put:
      tags: [Users]
      summary: 5. Cập nhật thông tin người dùng (Update)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        200:
          description: Cập nhật thành công

    delete:
      tags: [Users]
      summary: 6. Xóa người dùng (Delete)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Xóa thành công

  # ===========================
  # GROUP 2: PRODUCTS (5 APIs)
  # ===========================
...
  /product/cau10/{price}/{quantity}:
    get:
      tags: [Queries]
      summary: 10. Tìm theo điều kiện Giá < X và Số lượng > Y
      parameters:
        - in: path
          name: price
          required: true
          schema:
            type: number
          description: Mức giá trần (X)
        - in: path
          name: quantity
          required: true
          schema:
            type: number
          description: Mức số lượng sàn (Y)
      responses:
        200:
          description: Thành công

  # Câu 11: Giá < 100k, status true, sort giá giảm
...
  /product/cau15/{cateID}:
    get:
      tags: [Queries]
      summary: 15. Lấy theo Category ID (chỉ lấy Active)
      parameters:
        - in: path
          name: cateID
          required: true
          schema:
            type: string
          description: ID của danh mục
      responses:
        200:
          description: Thành công